test: move to functional
